// apps/web/generated/api/models/projectStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  draft: 'draft',
  published: 'published',
  collecting_applications: 'collecting_applications',
  assigned: 'assigned',
  in_progress: 'in_progress',
  delivery_submitted: 'delivery_submitted',
  under_supervisor_review: 'under_supervisor_review',
  revision_requested: 'revision_requested',
  awaiting_customer_review: 'awaiting_customer_review',
  completed: 'completed',
  cancelled: 'cancelled',
  archived: 'archived',
} as const;