// apps/web/generated/api/models/budgetResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetType } from './budgetType';

export interface BudgetResponse {
  budget_type: BudgetType;
  fixed_amount: string | null;
  min_amount: string | null;
  max_amount: string | null;
  currency_code: string;
}